requirements.txt
fastapi
sqlmodel
uvicorn
pika
psycopg2
python-dotenv
